refactor: Make better use of server components and suspense
