+++ src/components/Count/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Count.server.tsx";
